Fashion_Bot/pyproject.toml
[project]
name = "Fashion_Bot"
version = "0.1.0"
description = "Telegram-бот для подбора образов"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.13.1",
    "celery[redis]>=5.6.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["Fashion_Bot"]

